+++ test/support/test_repo.ex
@@ -1,0 +1,4 @@
+defmodule AshAge.TestRepo do
+  @moduledoc false
+  use Ecto.Repo, otp_app: :ash_age, adapter: Ecto.Adapters.Postgres
+end